feat: Add more command options & clean up
Add a fancy new way to create commands. Well... not really "new", but it
gives you a new location for it so that everything doesn't have to be
defined in one file. Also an example of how to do things by defining
things inside of a module.

Add a stock ticker command so you know how much money you've lost today.

General cleanup of some of the code.

A couple test examples.